+++ bgw-examples/bgw-docs-examples/src/main/kotlin/examples/components/container/ContainerExample.kt
@@ -6,7 +6,6 @@ import tools.aqua.bgw.components.container.LinearLayout
 import tools.aqua.bgw.components.container.Satchel
 import tools.aqua.bgw.components.gamecomponentviews.CardView
 import tools.aqua.bgw.components.gamecomponentviews.TokenView
-import tools.aqua.bgw.components.layoutviews.GridPane
 import tools.aqua.bgw.components.layoutviews.Pane
 import tools.aqua.bgw.components.uicomponents.Button
 import tools.aqua.bgw.components.uicomponents.Label
@@ -16,7 +15,6 @@ import tools.aqua.bgw.core.*
 import tools.aqua.bgw.util.Font
 import tools.aqua.bgw.visual.ColorVisual
 import tools.aqua.bgw.visual.ImageVisual
-import javax.swing.text.ComponentView
 
 fun main() {
 	ContainerExample()