+++ bgw-examples/bgw-tetris-example/src/main/kotlin/tools/aqua/bgw/examples/tetris/view/Refreshable.kt
@@ -1,0 +1,9 @@
+package tools.aqua.bgw.examples.tetris.view
+
+import tools.aqua.bgw.examples.tetris.entity.Tetris
+
+interface Refreshable {
+	fun refresh(tetris : Tetris)
+	fun hideStartInstructions()
+	fun loose()
+}