+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+[0.5]: https://github.com/tudo-aqua/bgw/releases/tag/v0.5
 [0.4]: https://github.com/tudo-aqua/bgw/releases/tag/v0.4
 [0.3]: https://github.com/tudo-aqua/bgw/releases/tag/v0.3
 [0.2]: https://github.com/tudo-aqua/bgw/releases/tag/v0.2
@@ -12,7 +13,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.5] - To be released
+
+### Added
+- ``ParallelAnimation`` and ``SequentialAnimation``.
+- Scene wide key events.
+- Sudoku example.
+
 ## [0.4] - 22. Sep. 2021
+
 ### Fixed
 - Empty grid columns and rows no longer get rendered size 0.0 in case of fixed dimensions.
 - Race condition while changing GameScenes caused by slow renderer.